fix flaky test_user_relationship_test.exs:81
It has an inserted_at which updates.
This caused the test to be flaky.
The implementation doesn't say inserted_at shouldn't update, so I changed the tests to match the current implementation.
It tests if the id remains the same and I check that there is indeed only one record with this id.

I did try to exclude the inserted_at to update, but that doesn't work.
I'm unsure if that's a bug in ecto or expected behaviour.
But since this wasn't part of the implementation in the first place, I assume it's ok to have it as it is now.