Set uploader to Pleroma.Uploaders.Local for failing tests
There were tests who were failing with

```
     The following arguments were given to IO.chardata_to_string/1:

         # 1
         nil

     Attempted function clauses (showing 2 out of 2):

         def chardata_to_string(string) when is_binary(string)
         def chardata_to_string(list) when is_list(list)
```

I noticed for several of them the problem was that the uploader was set to S3, while the rest of the S3 settings were nil.
The uploader itself didn't really matter much for the tests, but if it throws errors because it has settings that can't work, then the test fails.
I made changed the uploader to local first.