Fix flaky test_user_relationship_test.exs:81
Fix test so it works with just mocking `NaiveDateTime.utc_now/0` while having a passthroug hfor everything else

I also set async to false because AFAICT the tests can't run async with Mock.
When I didn't have the passthrough yet, I saw an unrelated test fail with `function NaiveDateTime.from_erl!/2 is undefined or private`.
This was a function I mocked with specific parameters. So this seems to imply that the mock was active on this other test.
The docs don't talk about being able to run async either.