Mock `NaiveDateTime.utc_now/0` instead of using 5 seconds tolerance
In filter_controller_test.exs:200 a 5 seconds tolerance was previously used to reduce flakyness of tests.
`NaiveDateTime.utc_now/0` is now mocked so we can always check on specific timing instead of an aproximate one.
async is set to false because I previously noticed that this Mock interferes with concurent tests.