Add mix task to send a notifaction, with a web push
Sending only a Push notification can't be done, it needs a Notification.
Creating a notification needs an activity.
Creating an activity needs an object.
That's why it's running throughout the whole process.

It isn't really clean because it's forging a post, it might be
acceptable to perform it from an account like "admin" or similar.

That said, you can't use Pleroma's builtin users ('relay' and
'internal.fetch') because they are invisible and stripped out from
notification creation pipeline.
It would have been a better solution, to have real administrative users.