add activity factories for undo and delete, fix remove factory
the undo activity factory creates a note, likes it, and then
creates an activity to undo the like.

the delete activity factory creates a note, and then creates
an activity to delete it.

remove_activity_factory's function name was misspelled, and
it does not appear to have been used anywhere.